FE-1327 - Move hooks to a common file, refactor to re-use some logic where possible
