+++ src/hooks/api/apiHooks.stories.mdx
@@ -1,6 +1,6 @@
 import { Meta, Story, Preview, Props } from '@storybook/addon-docs/blocks';
 
-<Meta title="Hooks|useSparkPostQuery" />
+<Meta title="Hooks|API" />
 
 `useSparkPostQuery` is a wrapper for React Query's
 [useQuery](https://react-query.tanstack.com/docs/api#usequery) hook that adds some additional error
