+++ src/hooks/index.js
@@ -9,6 +9,6 @@ export { default as usePinnedReport } from './usePinnedRport';
 export { default as usePrepareReportBuilderQuery } from './usePrepareReportBuilderQuery';
 export { default as usePrevious } from './usePrevious';
 export { default as useProvidedContext } from './useProvidedContext';
-export { useSparkPostQueries, useSparkPostQuery } from './useSparkPostQuery';
 export { default as useTabs } from './useTabs';
 export { default as useUniqueId } from './useUniqueId';
+export { useSparkPostQueries, useSparkPostQuery } from './api';
