--- src/hooks/useSparkPostQuery/index.js
@@ -1,1 +1,0 @@
-export { useSparkPostQuery, useSparkPostQueries } from './useSparkPostQuery';
